+++ common/src/main/java/com/gitlab/srcmc/rctmod/api/data/save/TrainerPlayerData.java
@@ -28,9 +28,11 @@ import com.gitlab.srcmc.rctmod.ModCommon;
 import com.gitlab.srcmc.rctmod.api.RCTMod;
 import com.gitlab.srcmc.rctmod.api.data.sync.PlayerState;
 import com.gitlab.srcmc.rctmod.api.utils.ChatUtils;
+import com.gitlab.srcmc.rctmod.api.utils.LangKeys;
 
 import net.minecraft.core.HolderLookup.Provider;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.saveddata.SavedData;
 
@@ -189,7 +191,7 @@ public class TrainerPlayerData extends SavedData {
         if(this.testSeriesCompleted()) {
             this.addSeriesCompletion(this.getCurrentSeries());
             this.updateLevelCap();
-            ChatUtils.sendTitle(this.player, "Completed", RCTMod.getInstance().getSeriesManager().getGraph(this.getCurrentSeries()).getMetaData().title());
+            ChatUtils.sendTitle(this.player, Component.translatable(LangKeys.GUI_TITLE_SERIES_COMPLETED).getString(), RCTMod.getInstance().getSeriesManager().getGraph(this.getCurrentSeries()).getMetaData().title().asComponent().getString());
             // ModRegistries.CriteriaTriggers.DEFEAT_COUNT.get().trigger((ServerPlayer)player, mob); // TODO: series completion advancements?
         }
     }